pyproject.toml
[project]
name = "skyla-prep"
version = "0.1.0"
description = "Credit Union Data Pipeline — PII scrubbing + Claude analytics"
requires-python = ">=3.10"
dependencies = [
    "faker>=28.0",
    "presidio-analyzer>=2.2",
    "presidio-anonymizer>=2.2",
    "spacy>=3.7",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "anthropic>=0.39",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "matplotlib>=3.8",
]

[project.optional-dependencies]
dev = ["httpx", "pytest"]
